move common code to utils
